% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix}
\source{
https://lukaswallrich.github.io/timesaveR/
}
\usage{
cor_matrix(x)
}
\arguments{
\item{x}{A data frame with numeric columns.}
}
\value{
A correlation matrix object with additional attributes for
formatting.
}
\description{
Creates a correlation matrix from numeric data. This is a simplified
replacement for timesaveR::cor_matrix().
}
\keyword{internal}
