% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divMetrics.R
\docType{package}
\name{divMetrics-package}
\alias{divMetrics}
\alias{divMetrics-package}
\title{divMetrics: Computation and Visual Exploration of Diversity Indices}
\description{
The divMetrics package provides functions for calculating a range of diversity
measures and for visualizing them. It enables users to explore divergences
between diversity measures to pick one that matches their research questions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lukaswallrich.github.io/divMetrics/}
  \item \url{https://github.com/LukasWallrich/divMetrics}
  \item Report bugs at \url{https://github.com/LukasWallrich/divMetrics/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{l.wallrich@bbk.ac.uk}

}
\keyword{internal}
