% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variety.R
\name{compute_Blau}
\alias{compute_Blau}
\title{Compute Blau's Index}
\usage{
compute_Blau(
  x,
  group = NULL,
  bin_width = NULL,
  bins = NULL,
  na.rm = FALSE,
  verbose = FALSE,
  return_df = FALSE
)
}
\arguments{
\item{x}{Numeric vector.}

\item{group}{Optional grouping variable of the same length as x.}

\item{bin_width}{Positive numeric value specifying bin width.}

\item{bins}{Optional numeric vector of bin boundaries. Values are assigned to bins using left-inclusive, right-exclusive intervals (i.e., first interval [a, b) does not include b).#' @param na.rm Logical. If TRUE, NA values are removed.}

\item{verbose}{Logical. If TRUE, prints the bin breaks.}

\item{return_df}{Logical. If TRUE, returns a dataframe with group, group_members and index_value.}
}
\value{
A named numeric vector with one value per group if group is provided (or a single value), unless return_df is \code{TRUE}.
}
\description{
Computes Blau's Index for a numeric vector, optionally within groups. Continuous values
can be binned either by specifying a bin width or explicit bin boundaries.
}
