% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{compute_CEI}
\alias{compute_CEI}
\title{Compute the Coverage and Evenness Index (CEI) or its Components}
\usage{
compute_CEI(
  x,
  group = NULL,
  range = NULL,
  na.rm = FALSE,
  return = "CEI",
  verbose = FALSE,
  return_df = FALSE
)
}
\arguments{
\item{x}{Numeric vector.}

\item{group}{Optional grouping variable of the same length as x.}

\item{range}{Numeric vector of length 2 specifying the theoretical min and max; defaults to the observed range.}

\item{na.rm}{Logical. If TRUE, NA values are removed.}

\item{return}{Character: "CEI" (default), "S", or "E".}

\item{verbose}{Logical. If TRUE, prints the range used.}

\item{return_df}{Logical. If TRUE, returns a dataframe with group, group_members and index_value.}
}
\value{
A named numeric vector with one value per group if group is provided (or a single value), unless return_df is \code{TRUE}.
}
\description{
Computes the Coverage and Evenness Index (CEI), or its components: Coverage (Coverage) and
Evenness Factor (E), for a numeric vector, optionally within groups. C is the proportion
of the theoretical range covered by the data, and
E is based on the uniformity of the distribution of values.
}
