% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report_teams}
\alias{report_teams}
\title{Report Teams}
\usage{
report_teams(attribute, team)
}
\arguments{
\item{attribute}{A vector of attribute values.}

\item{team}{A vector indicating the team for each attribute.}
}
\value{
A named character vector where each element is a comma-separated string of attribute values for a team.
The names correspond to the teams.
}
\description{
Combines attribute values by team, maintaining the order of first appearance.
}
\examples{
report_teams(c("A", "B", "C", "A", "A", "A"), c(1,1,1,2,2,2))
report_teams(c(1,2,3,4), c(2,2,1,1))
}
